Allow kubeproxy configuration in kind (to support nftables) (#1854)
* Handle default Submariner configuration

The top-level Submariner configuration overwrites any per-cluster
configuration; this fixes that to match the configuration handling for
other fields.

Signed-off-by: Stephen Kitt <skitt@redhat.com>

* Provide kubeproxy mode configuration in settings files

This adds support for configuring kubeproxy mode in settings files:
.kubeProxyMode is read from the cluster nodes, from the root node, and
defaults to iptables for now. Supported values in kind are iptables,
nftables, and ipvs.

Signed-off-by: Stephen Kitt <skitt@redhat.com>

* Handle nftables configuration with using

This allows nftables to be set by default with "using=nftables".

Signed-off-by: Stephen Kitt <skitt@redhat.com>

---------

Signed-off-by: Stephen Kitt <skitt@redhat.com>
Co-authored-by: Yossi Boaron <yboaron@redhat.com>